+++ jupiter-docker/src/test/java/no/mnemonic/commons/jupiter/docker/CassandraDockerExtensionTest.java
@@ -66,9 +66,7 @@ public class CassandraDockerExtensionTest {
   public void testIsContainerReachableFailsOnExecCreateException() throws Throwable {
     mockStartContainer();
     when(dockerClient.execCreate(any(), any(), any())).thenThrow(DockerException.class);
-
-    IllegalStateException ex = assertThrows(IllegalStateException.class, () -> extension.beforeAll(null));
-    assertTrue(ex.getMessage().contains("Could not execute 'cqlsh'"));
+    assertThrows(TimeoutException.class, () -> extension.beforeAll(null));
   }
 
   @Test
@@ -76,9 +74,7 @@ public class CassandraDockerExtensionTest {
     mockStartContainer();
     when(dockerClient.execCreate(any(), any(), any())).thenReturn(execCreation("executionID"));
     when(dockerClient.execStart("executionID")).thenThrow(DockerException.class);
-
-    IllegalStateException ex = assertThrows(IllegalStateException.class, () -> extension.beforeAll(null));
-    assertTrue(ex.getMessage().contains("Could not execute 'cqlsh'"));
+    assertThrows(TimeoutException.class, () -> extension.beforeAll(null));
   }
 
   @Test